Remove the check that profitable priority fee must be greater than minGasPrice
